Move from ros_ign_utils to ros_ign_gazebo
