--- ros_ign_utils/CMakeLists.txt
@@ -1,62 +1,0 @@
-cmake_minimum_required(VERSION 3.10.2 FATAL_ERROR)
-
-project(ros_ign_utils)
-
-find_package(ament_cmake REQUIRED)
-
-find_package(rclcpp REQUIRED)
-find_package(rclcpp_components REQUIRED)
-find_package(geometry_msgs REQUIRED)
-find_package(sensor_msgs REQUIRED)
-find_package(tf2 REQUIRED)
-find_package(tf2_ros REQUIRED)
-
-add_library(${PROJECT_NAME}
-  SHARED
-  src/optical_frame_publisher.cc
-)
-
-target_link_libraries(
-  ${PROJECT_NAME}
-  ${geometry_msgs_TARGETS}
-  rclcpp::rclcpp
-  rclcpp_components::component
-  ${sensor_msgs_TARGETS}
-  tf2::tf2
-  tf2_ros::tf2_ros
-)
-
-target_include_directories(${PROJECT_NAME} PUBLIC
-  "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
-  "$<INSTALL_INTERFACE:include/${PROJECT_NAME}>")
-
-rclcpp_components_register_node(${PROJECT_NAME}
-  PLUGIN "ros_ign_utils::OpticalFramePublisher"
-  EXECUTABLE optical_frame_publisher
-)
-
-install(TARGETS ${PROJECT_NAME}
-  EXPORT export_${PROJECT_NAME}
-  ARCHIVE DESTINATION lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin)
-
-install(DIRECTORY include/
-  DESTINATION include/${PROJECT_NAME})
-
-if(BUILD_TESTING)
-  # Disable some broken tests for now
-  set(ament_cmake_copyright_FOUND TRUE)
-
-  find_package(ament_lint_auto REQUIRED)
-  ament_lint_auto_find_test_dependencies()
-endif()
-
-# Export old-style CMake variables
-ament_export_include_directories("include/${PROJECT_NAME}")
-ament_export_libraries(${PROJECT_NAME})
-
-# Export modern CMake targets
-ament_export_targets(export_${PROJECT_NAME})
-
-ament_package()
