--- ros_ign_utils/package.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>ros_ign_utils</name>
-  <version>0.0.0</version>
-  <description>Utility functions and applications for Gazebo ROS interactions</description>
-
-  <maintainer email="michael@openrobotics.org">Michael Carroll</maintainer>
-
-  <license>Apache License 2.0</license>
-
-  <author>Michael Carroll</author>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
-
-  <test_depend>ament_cmake_gtest</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-
-</package>
-
